+++ frontend/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "axios": "^0.25.0",
         "core-js": "^3.6.5",
+        "normalize.css": "^8.0.1",
         "vue": "^3.0.0",
         "vue-router": "^4.0.0-0"
     },
@@ -30,7 +31,6 @@
         "eslint": "^6.7.2",
         "eslint-plugin-vue": "^7.0.0",
         "jest": "^26.6.3",
-        "normalize.css": "^8.0.1",
         "ts-jest": "^26.5.0",
         "typescript": "~3.9.3",
         "vue-jest": "^5.0.0-0"
