--- test.js
@@ -1,16 +1,0 @@
-import {
-    shallowMount
-} from '@vue/test-utils'
-import HelloWorld from '@/components/HelloWorld.vue'
-
-describe('HelloWorld.vue', () => {
-    it('renders props.msg when passed', () => {
-        const msg = 'new message'
-        const wrapper = shallowMount(HelloWorld, {
-            props: {
-                msg
-            }
-        })
-        expect(wrapper.text()).toMatch(msg)
-    })
-})
